# [Next] Use dynamic array while evaluating circuit