Right know, we keep a vector of MLIR values while evaluating a circuit (one value per gate). This implies that we are using a lot of MLIR values at the same time, which puts a lot of pressure on the LLVM optimization phases.

This PR changes it to use a dynamic array of gate values, which uses a single register (a pointer to the array).

# Benchmark

I compiled several circuit-heavy contract classes, and compared the compilation time before and this PR:

| Class Hash | Before | After |
| ---------- | ------ | ----- |
| 0x1b5fbe104c033025dbb7fb37011781cc9344e881b4828cdaa023a80fecafde4 | 77024 ms | 91510 ms | 
| 0x3100defca27214e5f78f25e48a5b05e45899c6834cb4d34f48384c18e14dff7 | 74235 ms | 88951 ms | 
| 0x79666cdb4fc3cbcafbd74f4ea4e2855bf455c5a7c70915f5679325c54032771 | 87430 ms | 88287 ms | 
| 0xa40ecaa08eaef629d433b15f6b017461df8bea9b4e7299eb8ba3632e32b5a5 | 96162 ms | 97812 ms | 
| 0x4ffeed293927cd56686a9038a10026a2d3b9602f789d1f163c1c4ac9a822a82 | 107289 ms | 129285 ms | 
| 0x2269858a40ea0535cb373b0c981c91b907466edf6d65bcaf669760bbee0ae4d | 98239 ms | 122565 ms | 
| 0x5ff378cb2f16804539ecb92e84f273aafbab57d450530e9fe8e87771705a673 | 88499 ms | 100990 ms | 
| 0x4edde37ca59d9dff8f4ac8945b1c4860b606abd61d74727904ad7494fccdfa9 | 88920 ms | 101077 ms | 

## Checklist
- [ ] Linked to Github Issue
- [ ] Unit tests added
- [ ] Integration tests added.
- [ ] This change requires new documentation.
  - [ ] Documentation has been added/updated.