@gabrielbosio 2025-08-05
Both `Hyperfine` and `Compare Dumps` are not running since get triggered on a PR to `main` only.